Parsing is a cancellable operation; Includes unit tests to test parsing cancellation.
The callbacks provided to the CSVParserConfiguration now return a "ParsingDisposition" enum to indicate whether parsing should continue or should be cancelled. Unimplemented callbacks are assumed to return "Continue". A callback return value was chosen over providing some sort of cancel method, cancel block, etc, because a cancellation block could be invoked at inappropriate times. By using a return value, we reduce the amount of mutable parsing state that needs to be kept and can use that to enforce parsing correctness.